:art: restructured JSONParser and started with Decoder.
