+++ src/main/code/de/tuchemnitz/se/exercise/core/configmanager/ConfigFileDecoder.kt
@@ -1,0 +1,7 @@
+package de.tuchemnitz.se.exercise.core.configmanager
+
+import java.nio.file.Path
+
+class ConfigFileDecoder(configPath: Path) {
+    var configFileContents = ConfigFile
+}
