+++ src/main/code/de/tuchemnitz/se/exercise/persist/configs/EyeTrackingConfig.kt
@@ -1,5 +1,6 @@
 package de.tuchemnitz.se.exercise.persist.configs
 
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import org.litote.kmongo.Id
