+++ src/main/code/de/tuchemnitz/se/exercise/persist/configs/ZoomMapsConfig.kt
@@ -1,6 +1,7 @@
 package de.tuchemnitz.se.exercise.persist.configs
 
 import de.tuchemnitz.se.exercise.persist.now
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import org.litote.kmongo.Id
