+++ src/test/code/de/tuchemnitz/se/exercise/core/configmanager/ConfigManagerTest.kt
@@ -6,16 +6,12 @@ import assertk.assertions.isNull
 import com.mongodb.client.FindIterable
 import de.tuchemnitz.se.exercise.DummyData
 import de.tuchemnitz.se.exercise.persist.configs.CodeChartsConfig
-import de.tuchemnitz.se.exercise.persist.configs.EyeTrackingConfig
 import de.tuchemnitz.se.exercise.persist.configs.collections.CodeChartsConfigCollection
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import org.bson.BsonDocument
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.params.ParameterizedTest
@@ -97,51 +93,5 @@ class ConfigManagerTest {
 
     @Test
     fun `invoking of saving functions of config should work`() {
-    }
-
-    @Nested
-    inner class DataBaseUsage {
-
-        @BeforeEach
-        fun setup() {
-            DummyData.codeChartsConfigs.forEach {
-                DummyData.codeChartsConfigCollection.saveOne(it)
-            }
-
-            DummyData.zoomMapsConfigs().forEach {
-                DummyData.zoomMapsConfigCollection.saveOne(it)
-            }
-        }
-
-        @AfterEach
-        fun tearDown() {
-            DummyData.codeChartsConfigCollection.deleteMany()
-            DummyData.zoomMapsConfigCollection.deleteMany()
-        }
-
-        @Test
-        fun `assembling all database configs should work`() { // integration test
-            val expected = ToolConfigs(
-                codeChartsConfig = mostRecentCodeChartsConfig,
-                zoomMapsConfig = mostRecentZoomMapsConfig,
-                // TODO
-                eyeTrackingConfig = EyeTrackingConfig(dummyVal = ""),
-                // TODO
-                bubbleViewConfig = BubbleViewConfig(filter = setOf(0F))
-            )
-            val actual = DummyData.manager.assembleAllConfigurations()
-                .copy(
-                    eyeTrackingConfig = expected.eyeTrackingConfig,
-                    bubbleViewConfig = expected.bubbleViewConfig
-                )
-            assertThat(actual).isEqualTo(expected)
-            DummyData.manager.writeFile(Path.of("test.cfg"), DummyData.manager.configFile)
-        }
-
-        private val mostRecentCodeChartsConfig =
-            DummyData.codeChartsConfigs.maxByOrNull { it.savedAt }
-
-        private val mostRecentZoomMapsConfig =
-            DummyData.zoomMapsConfigs.maxByOrNull { it.savedAt }
     }
 }
